requirements.txt
fastapi
uvicorn
langchain
pymongo
python-dotenv
httpx
python-multipart
pydantic
langchain_groq
openai-whisper
vosk
sounddevice
openai